Change ids type to uint

+++ Giselle.Net.EtherNetIP/AbstractOptions.cs
@@ -20,7 +20,7 @@ namespace Giselle.Net.EtherNetIP
         /// </summary>
         public byte TimeoutMultiplier { get; set; }
 
-        public ushort ClassID { get; set; }
+        public uint ClassID { get; set; }
 
         public AbstractOptions()
         {
